refactor(payment): move some util functions from payment_utils.py
And adjust the imports
